docker-commit

+++ advert/app/utils/favoriteImages.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from 'axios';
-
+import Cookies from 'js-cookie';
 const API_BASE_URL = process.env.MY_APP_BASE_URL || 'http://localhost:3500';
 
 interface Product {
@@ -45,11 +45,15 @@ interface AuthApiResponse extends ApiResponse<any> {
 export const addToFavoriteImages = async (productId: string):Promise<ApiResponse<any>> => {
   try {
 
-    const userId = localStorage.getItem('userId');
-    
-    const token = localStorage.getItem('accessToken');
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<AuthApiResponse> = await axios.post(
@@ -73,10 +77,16 @@ export const addToFavoriteImages = async (productId: string):Promise<ApiResponse
 
 export const getFavoriteImages = async (): Promise<ApiResponse<Product[]>> => {
   try {
-    const userId = localStorage.getItem('userId');
-    const token = localStorage.getItem('accessToken');
+
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse<Product[]>> = await axios.get(
@@ -97,10 +107,16 @@ export const getFavoriteImages = async (): Promise<ApiResponse<Product[]>> => {
 
 export const removeFavoriteImages = async (favoriteId: string): Promise<ApiResponse<any>> => {
   try {
-    const userId = localStorage.getItem('userId');
-    const token = localStorage.getItem('accessToken');
+
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse<any>> = await axios.delete(
