+++ advert/app/utils/favoriteVideo.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from 'axios';
-
+import Cookies from 'js-cookie';
 const API_BASE_URL = process.env.MY_APP_BASE_URL || 'http://localhost:3500';
 
 interface Product {
@@ -45,11 +45,15 @@ interface AuthApiResponse extends ApiResponse<any> {
 export const addToFavoriteVideo = async (productId: string): Promise<ApiResponse<any>> => {
   try {
 
-    const userId = localStorage.getItem('userId');
-    
-    const token = localStorage.getItem('accessToken');
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<AuthApiResponse> = await axios.post(
@@ -72,10 +76,16 @@ export const addToFavoriteVideo = async (productId: string): Promise<ApiResponse
 
 export const getFavoriteVideo = async ():  Promise<ApiResponse<Product[]>> => {
   try {
-    const userId = localStorage.getItem('userId');
-    const token = localStorage.getItem('accessToken');
+
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse<Product[]>> = await axios.get(
@@ -96,10 +106,16 @@ export const getFavoriteVideo = async ():  Promise<ApiResponse<Product[]>> => {
 
 export const removeFavoriteVideo = async (favoriteId: string): Promise<ApiResponse<any>> => {
   try {
-    const userId = localStorage.getItem('userId');
-    const token = localStorage.getItem('accessToken');
+
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse<any>> = await axios.delete(
