+++ advert/app/utils/profile.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from 'axios';
-
+import Cookies from 'js-cookie';
 const API_BASE_URL = process.env.MY_APP_BASE_URL || 'http://localhost:3500';
 
 interface ApiResponse { 
@@ -18,11 +18,15 @@ interface AuthApiResponse extends ApiResponse {
 export const uploadProfileImage = async ( image: File): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
-    
-    const token = localStorage.getItem('accessToken');
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const formData = new FormData();
@@ -49,11 +53,15 @@ export const uploadProfileImage = async ( image: File): Promise<ApiResponse> => 
 export const getProfileImage = async (): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse> = await axios.get(
@@ -76,11 +84,15 @@ export const getProfileImage = async (): Promise<ApiResponse> => {
 export const updateProfileImage = async ( image: File): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const formData = new FormData();
@@ -107,11 +119,15 @@ export const updateProfileImage = async ( image: File): Promise<ApiResponse> => 
 export const deleteProfileImage = async (): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse> = await axios.delete(
