+++ advert/app/utils/videos.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from 'axios';
-
+import Cookies from 'js-cookie';
 const API_BASE_URL = process.env.MY_APP_BASE_URL || 'http://localhost:3500';
 
 interface Video {
@@ -54,10 +54,16 @@ interface AuthApiResponse extends ApiResponse {
 
 export const uploadVideo = async (formData: FormData): Promise<ApiResponse> => {
   try {
-    const userId = localStorage.getItem('userId');
-    const token = localStorage.getItem('accessToken');
+
+    const userId = Cookies.get('userId');
+
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<ApiResponse> = await axios.post(
@@ -75,9 +81,9 @@ export const uploadVideo = async (formData: FormData): Promise<ApiResponse> => {
     const videos = response.data.videos;
     if (videos) {
       const videoId = videos[0].id;
-      localStorage.setItem('videoId', videoId.toString());
+      Cookies.set('videoId', videoId.toString());
     }
-    
+   
     return response.data;
   } catch (error: any) {
     console.error('Error uploading video:', error.message);
@@ -88,11 +94,15 @@ export const uploadVideo = async (formData: FormData): Promise<ApiResponse> => {
 export const deleteVideo = async (productId:number): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<AuthApiResponse> = await axios.delete(
@@ -115,11 +125,15 @@ export const deleteVideo = async (productId:number): Promise<ApiResponse> => {
 export const editVideo = async (videoId: number): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     const response: AxiosResponse<AuthApiResponse> = await axios.put(
@@ -143,11 +157,15 @@ export const editVideo = async (videoId: number): Promise<ApiResponse> => {
 export const getVideo = async (filter: FilterState, searchQuery?:string): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('userId');
+    const userId = Cookies.get('userId');
 
-    const token = localStorage.getItem('accessToken');
+    if (!userId) {
+      throw new Error('User ID not found in cookies');
+    }
+
+    const token = Cookies.get('accessToken');
     if (!token) {
-      throw new Error('Access token not found');
+      throw new Error('Access token not found in cookies');
     }
 
     let apiUrl = `${API_BASE_URL}/videos/getVideo?city=${filter.city}&country=${filter.country}&state=${filter.state}&fullAddress=${filter.fullAddress}&minPrice=${filter.minPrice}&maxPrice=${filter.maxPrice}&company=${filter.company}&name=${filter.name}&types=${filter.types}&categories=${filter.categories}`;
@@ -177,7 +195,7 @@ export const getVideo = async (filter: FilterState, searchQuery?:string): Promis
 export const getVideoByProfileId = async (): Promise<ApiResponse> => {
   try {
 
-    const userId = localStorage.getItem('profileId');
+    const userId = Cookies.get('profileId');
 
     const response: AxiosResponse<AuthApiResponse> = await axios.get(`${API_BASE_URL}/videos/getVideosByProfileId`,
     
@@ -225,7 +243,7 @@ export const getVideos = async (filter: FilterState, searchQuery?:string): Promi
 export const getVideosById = async (): Promise<ApiResponse> => {
   try {
 
-    const productId = localStorage.getItem('productId');
+    const productId = Cookies.get('productId');
 
     const response: AxiosResponse<AuthApiResponse> = await axios.get(`${API_BASE_URL}/videos/getVideosById`,
